Refactor: place api function call in App.js component Refactor: Missions to loop through the lists and return a mission Add: mission component to render individual mission
