+++ src/components/App.css
@@ -14,8 +14,8 @@ body {
 }
 
 .px-20 {
-  padding-right: 20%;
-  padding-left: 20%;
+  padding-right: 10%;
+  padding-left: 10%;
 }
 
 .py-30 {
