+++ src/components/Missions.js
@@ -1,21 +1,12 @@
 import React, { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { Table, Badge, Button } from 'react-bootstrap';
-import getMissionsFromApi from '../redux/api';
-import { getMissionsAction } from '../redux/missions';
+import { useSelector } from 'react-redux';
+import { Table } from 'react-bootstrap';
+import Mission from './Mission';
 
 export default function Missions() {
-  const dispatch = useDispatch();
+  const state = useSelector((state) => state.missionsReducer);
 
-  useEffect(() => {
-    getMissionsFromApi()
-      .then((data) => {
-        dispatch(getMissionsAction(data));
-      })
-      .catch((err) => {
-        throw new Error(err);
-      });
-  }, []);
+  useEffect(() => {}, []);
 
   return (
     <div className="px-20 py-30">
@@ -35,18 +26,11 @@ export default function Missions() {
           </tr>
         </thead>
         <tbody>
-          <tr className="text-lg">
-            <td className="font-bold">Mark</td>
-            <td>Otto</td>
-            <td><Badge bg="secondary">NOT A MEMBER</Badge></td>
-            <td><Button variant="outline-dark">Join Mission</Button></td>
-          </tr>
-          <tr className="text-lg">
-            <td className="font-bold">Jacob</td>
-            <td>Thornton</td>
-            <td><Badge bg="primary">ACTIVE MEMBER</Badge></td>
-            <td><Button variant="outline-danger">Leave Mission</Button></td>
-          </tr>
+          {
+            state.missionsArray.map(
+              (mission) => <Mission key={mission.missionId} detail={mission} />,
+            )
+          }
         </tbody>
       </Table>
     </div>
